requirements.txt
fastapi>=0.54.1,<0.55
locustio>=0.14.5,<0.15
uvicorn>=0.11.3,<0.12